Handled unusual case where arcgis.com doesn't create thumbnail for item
1. From My Content, add a zipped shapefile item from the computer. It
will not have even the default thumbnail.

2. Share the created item with the map gallery.

3. If you also make it public to everyone, you can look at its
definition using the URL
http://www.arcgis.com/sharing/content/items/<itemId>?f=pjson, where
<itemId> is the unique id that arcgis.com created for the new item.
You'll see that for this particular case,
"thumbnail": null,

4. If you display the gallery in the app, you should see the
graphics/missingThumbnail.png image as both the thumbnail (Home &
Gallery pages) and the blown-up view.

5. Go back to the shapefile item and edit it.  If you give it a 200x133
pixel image, the thumbnails will be fine, but the blown-up view will be
fuzzy.  If you give it a bigger image such as 800x532 (4x larger),
arcgis.com will construct a thumbnail and save the bigger image.